feat: implement Frame class
Create the frame class. For each frame the bowler gets up to 2 tries to
knock down all the pins.

For this type of game, the max number of pins is 10.

* In each frame, the bowler gets up to two tries to knock down all the
pins.

* If in two tries, he fails to knock them all down, his score for that
frame is the total number of pins knocked down in his two tries.

* If in two tries he knocks them all down, this is called a “spare”
and his score for the frame is ten plus the number of pins knocked down
on his next throw (in his next turn).

* If on his first try in the frame he knocks down all the pins, this is
called a “strike”. His turn is over, and his score for the frame is ten
plus the simple total of the pins knocked down in his next two rolls.